Move error from utils to source
