+++ connectors/sources/mysql.py
@@ -14,14 +14,9 @@ from connectors.filtering.validation import (
     SyncRuleValidationResult,
 )
 from connectors.logger import logger
-from connectors.source import BaseDataSource
+from connectors.source import BaseDataSource, ConfigurableFieldValueError
 from connectors.sources.generic_database import WILDCARD, configured_tables, is_wildcard
-from connectors.utils import (
-    CancellableSleeps,
-    ConfigurableFieldValueError,
-    RetryStrategy,
-    retryable,
-)
+from connectors.utils import CancellableSleeps, RetryStrategy, retryable
 
 MAX_POOL_SIZE = 10
 QUERIES = {
