+++ connectors/sources/sharepoint.py
@@ -18,11 +18,10 @@ from aiofiles.tempfile import NamedTemporaryFile
 from aiohttp.client_exceptions import ClientResponseError, ServerDisconnectedError
 
 from connectors.logger import logger
-from connectors.source import BaseDataSource
+from connectors.source import BaseDataSource, ConfigurableFieldValueError
 from connectors.utils import (
     TIKA_SUPPORTED_FILETYPES,
     CancellableSleeps,
-    ConfigurableFieldValueError,
     RetryStrategy,
     convert_to_b64,
     evaluate_timedelta,
