+++ connectors/sources/tests/test_s3.py
@@ -10,9 +10,9 @@ from unittest import mock
 import aioboto3
 import pytest
 
+from connectors.source import ConfigurableFieldValueError
 from connectors.sources.s3 import S3DataSource
 from connectors.sources.tests.support import assert_basics, create_source
-from connectors.utils import ConfigurableFieldValueError
 
 
 @pytest.fixture(scope="session", autouse=True)
