+++ connectors/utils.py
@@ -472,7 +472,3 @@ def get_pem_format(key, max_split=-1):
     key = " ".join(key.split("\n", max_split))
     key = " ".join(key.rsplit("\n", max_split))
     return key
-
-
-class ConfigurableFieldValueError(Exception):
-    pass
